move nrf24l01 code to devices
