+++ board/Sipeed_LonganNano/eclipse/nRF24L01/main.c
@@ -1,45 +1,10 @@
 #include "mcu_init.h"
 #include "tos_k.h"
 #include "lcd.h"
-#include "nrf24.h"
-
 
 #define LCD_TASK_SIZE 1024
 k_task_t lcd_handle;
 uint8_t lcd_stk[LCD_TASK_SIZE];
-
-
-#define LED_TASK_SIZE 1024
-k_task_t led_handle;
-uint8_t led_stk[LED_TASK_SIZE];
-
-k_sem_t sem_led;
-
-typedef struct {
-    int port;
-    int pin;
-} Led_t;
-
-Led_t leds[] = {
-        { LEDR_GPIO_PORT, LEDR_PIN },
-        { LEDG_GPIO_PORT, LEDG_PIN },
-        { LEDB_GPIO_PORT, LEDB_PIN }
-};
-
-
-void task_led(void *arg)
-{
-    int task_cnt1 = 0;
-    while (1) {
-        //printf("hello world from %s cnt: %d\n", __func__, task_cnt1++);
-
-        tos_sem_pend(&sem_led, ~0);
-
-        gpio_bit_reset(LEDB_GPIO_PORT, LEDB_PIN);
-        tos_task_delay(50);
-        gpio_bit_set(LEDB_GPIO_PORT, LEDB_PIN);
-    }
-}
 
 
 void task_lcd(void *arg)
@@ -69,6 +34,8 @@ void task_lcd(void *arg)
     }
 }
 
+void nrf24l01_init();
+
 void main(void) {
     board_init();
 
@@ -76,13 +43,9 @@ void main(void) {
 
     tos_knl_init();
 
-
     nrf24l01_init();
 
-
-    tos_sem_create(&sem_led, 1);
-    tos_task_create(&led_handle,     "led", task_led, NULL, 6,   led_stk,   LED_TASK_SIZE, 0);
-    tos_task_create(&lcd_handle,     "lcd", task_lcd, NULL, 6,   lcd_stk,   LCD_TASK_SIZE, 0);
+    tos_task_create(&lcd_handle,     "lcd", task_lcd, NULL, 7,   lcd_stk,   LCD_TASK_SIZE, 0);
 
     tos_knl_start();
