--- board/Sipeed_LonganNano/eclipse/nRF24L01/nrf24.h
@@ -1,11 +1,0 @@
-#ifndef NRF24_H_
-#define NRF24_H_
-
-#include "stdlib.h"
-#include "gd32vf103_gpio.h"
-#include "nrf24l01_hal.h"
-
-void nrf24l01_init();
-
-
-#endif /* NRF24_H_ */
