--- board/TencentOS_tiny_EVB_LX/eclipse/nRF24L01/.gitignore
@@ -1,1 +1,0 @@
-/Debug/
