--- board/TencentOS_tiny_EVB_LX/eclipse/nRF24L01/.project
@@ -1,94 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<projectDescription>
-	<name>nRF24L01</name>
-	<comment></comment>
-	<projects>
-	</projects>
-	<buildSpec>
-		<buildCommand>
-			<name>org.eclipse.cdt.managedbuilder.core.genmakebuilder</name>
-			<triggers>clean,full,incremental,</triggers>
-			<arguments>
-			</arguments>
-		</buildCommand>
-		<buildCommand>
-			<name>org.eclipse.cdt.managedbuilder.core.ScannerConfigBuilder</name>
-			<triggers>full,incremental,</triggers>
-			<arguments>
-			</arguments>
-		</buildCommand>
-	</buildSpec>
-	<natures>
-		<nature>org.eclipse.cdt.core.cnature</nature>
-		<nature>org.eclipse.cdt.managedbuilder.core.managedBuildNature</nature>
-		<nature>org.eclipse.cdt.managedbuilder.core.ScannerConfigNature</nature>
-	</natures>
-	<linkedResources>
-		<link>
-			<name>Application</name>
-			<type>2</type>
-			<locationURI>virtual:/virtual</locationURI>
-		</link>
-		<link>
-			<name>GD32VF103_Firmware_Library</name>
-			<type>2</type>
-			<locationURI>$%7BPARENT-4-PROJECT_LOC%7D/platform/vendor_bsp/gd/GD32VF103_Firmware_Library</locationURI>
-		</link>
-		<link>
-			<name>TencentOS_tiny</name>
-			<type>2</type>
-			<locationURI>virtual:/virtual</locationURI>
-		</link>
-		<link>
-			<name>Application/Inc</name>
-			<type>2</type>
-			<locationURI>$%7BPARENT-2-PROJECT_LOC%7D/BSP/Inc</locationURI>
-		</link>
-		<link>
-			<name>Application/Src</name>
-			<type>2</type>
-			<locationURI>$%7BPARENT-2-PROJECT_LOC%7D/BSP/Src</locationURI>
-		</link>
-		<link>
-			<name>Application/tos_config.h</name>
-			<type>1</type>
-			<locationURI>$%7BPARENT-2-PROJECT_LOC%7D/TOS_CONFIG/tos_config.h</locationURI>
-		</link>
-		<link>
-			<name>TencentOS_tiny/arch</name>
-			<type>2</type>
-			<locationURI>virtual:/virtual</locationURI>
-		</link>
-		<link>
-			<name>TencentOS_tiny/kernel</name>
-			<type>2</type>
-			<locationURI>$%7BPARENT-4-PROJECT_LOC%7D/kernel</locationURI>
-		</link>
-		<link>
-			<name>TencentOS_tiny/arch/risc-v</name>
-			<type>2</type>
-			<locationURI>virtual:/virtual</locationURI>
-		</link>
-		<link>
-			<name>TencentOS_tiny/arch/risc-v/bumblebee</name>
-			<type>2</type>
-			<locationURI>TOP_DIR/arch/risc-v/bumblebee/gcc</locationURI>
-		</link>
-		<link>
-			<name>TencentOS_tiny/arch/risc-v/common</name>
-			<type>2</type>
-			<locationURI>$%7BPARENT-4-PROJECT_LOC%7D/arch/risc-v/common</locationURI>
-		</link>
-		<link>
-			<name>TencentOS_tiny/arch/risc-v/rv32i</name>
-			<type>2</type>
-			<locationURI>TOP_DIR/arch/risc-v/rv32i/gcc</locationURI>
-		</link>
-	</linkedResources>
-	<variableList>
-		<variable>
-			<name>TOP_DIR</name>
-			<value>$%7BPARENT-4-PROJECT_LOC%7D</value>
-		</variable>
-	</variableList>
-</projectDescription>
