--- board/TencentOS_tiny_EVB_LX/eclipse/nRF24L01/gd32vf103_libopt.h
@@ -1,61 +1,0 @@
-/*!
-    \file  gd32vf103_libopt.h
-    \brief library optional for gd32vf103
-
-    \version 2019-6-5, V1.0.0, demo for GD32VF103
-*/
-
-/*
-    Copyright (c) 2019, GigaDevice Semiconductor Inc.
-
-    Redistribution and use in source and binary forms, with or without modification, 
-are permitted provided that the following conditions are met:
-
-    1. Redistributions of source code must retain the above copyright notice, this 
-       list of conditions and the following disclaimer.
-    2. Redistributions in binary form must reproduce the above copyright notice, 
-       this list of conditions and the following disclaimer in the documentation 
-       and/or other materials provided with the distribution.
-    3. Neither the name of the copyright holder nor the names of its contributors 
-       may be used to endorse or promote products derived from this software without 
-       specific prior written permission.
-
-    THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" 
-AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED 
-WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. 
-IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, 
-INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT 
-NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR 
-PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, 
-WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) 
-ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY 
-OF SUCH DAMAGE.
-*/
-
-#ifndef GD32VF103_LIBOPT_H
-#define GD32VF103_LIBOPT_H
-
-#include "gd32vf103_adc.h"
-#include "gd32vf103_bkp.h"
-#include "gd32vf103_can.h"
-#include "gd32vf103_crc.h"
-#include "gd32vf103_dac.h"
-#include "gd32vf103_dma.h"
-#include "gd32vf103_eclic.h"
-#include "gd32vf103_exmc.h"
-#include "gd32vf103_exti.h"
-#include "gd32vf103_fmc.h"
-#include "gd32vf103_gpio.h"
-#include "gd32vf103_i2c.h"
-#include "gd32vf103_fwdgt.h"
-#include "gd32vf103_dbg.h"
-#include "gd32vf103_pmu.h"
-#include "gd32vf103_rcu.h"
-#include "gd32vf103_rtc.h"
-#include "gd32vf103_spi.h"
-#include "gd32vf103_timer.h"
-#include "gd32vf103_usart.h"
-#include "gd32vf103_wwdgt.h"
-#include "n200_func.h"
-
-#endif /* GD32VF103_LIBOPT_H */
