--- board/TencentOS_tiny_EVB_LX/eclipse/nRF24L01/openocd_gdlink.cfg
@@ -1,45 +1,0 @@
-adapter_khz     1000
-reset_config srst_only
-adapter_nsrst_assert_width 100
-
-
-
-interface cmsis-dap
-
-transport select jtag
-
-#autoexit true
-
-set _CHIPNAME riscv
-jtag newtap $_CHIPNAME cpu -irlen 5 -expected-id 0x1000563d
-
-set _TARGETNAME $_CHIPNAME.cpu
-target create $_TARGETNAME riscv -chain-position $_TARGETNAME
-$_TARGETNAME configure -work-area-phys 0x20000000 -work-area-size 20480 -work-area-backup 0
-
-
-# Work-area is a space in RAM used for flash programming
-if { [info exists WORKAREASIZE] } {
-   set _WORKAREASIZE $WORKAREASIZE
-} else {
-   set _WORKAREASIZE 0x5000
-}
-
-# Allow overriding the Flash bank size
-if { [info exists FLASH_SIZE] } {
-    set _FLASH_SIZE $FLASH_SIZE
-} else {
-    # autodetect size
-    set _FLASH_SIZE 0
-}
-
-# flash size will be probed
-set _FLASHNAME $_CHIPNAME.flash
-
-flash bank $_FLASHNAME gd32vf103 0x08000000 0 0 0 $_TARGETNAME
-riscv set_reset_timeout_sec 1
-init
-
-halt
-
-
